Создал Validation класс для валидации имени, логина, пароля
