--- app/src/main/kotlin/domain/Validate.kt
@@ -1,8 +1,0 @@
-package domain
-
-object Validate {
-    class InvalidNameException(message: String) : Exception(message)
-    class InvalidLoginException(message: String) : Exception(message)
-    class InvalidPasswordException(message: String) : Exception(message)
-    class NotEqualPasswordException(message: String) : Exception(message)
-}
