+++ app/src/main/kotlin/viewmodel/ViewModelProvider.kt
@@ -2,8 +2,8 @@ package viewmodel
 
 import auth.GrpcAuthenticatorStub
 import utils.LocalStorage
-
+import utils.Validator
 /**
  * Провайдер ViewModel для всех активностей
  */
-object ViewModelProvider: ViewModel by LoginViewModel(LocalStorage(), GrpcAuthenticatorStub())
+object ViewModelProvider: ViewModel by LoginViewModel(LocalStorage(), GrpcAuthenticatorStub(), Validator())
